refactor: SelectJdbcTemplate과 JdbcTemplate를 합친다.
