+++ app/src/main/java/com/techcourse/dao/JdbcTemplate.java
@@ -2,6 +2,7 @@ package com.techcourse.dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import javax.sql.DataSource;
 import org.slf4j.Logger;
@@ -17,6 +18,8 @@ public abstract class JdbcTemplate {
 
     protected abstract void setValues(PreparedStatement pstmt) throws SQLException;
 
+    protected abstract Object mapRow(ResultSet rs);
+
     public void update() throws SQLException {
         final String sql = createQuery();
 
@@ -27,6 +30,27 @@ public abstract class JdbcTemplate {
 
             setValues(pstmt);
             pstmt.executeUpdate();
+        }
+    }
+
+    private ResultSet executeQuery(PreparedStatement pstmt) throws SQLException {
+        return pstmt.executeQuery();
+    }
+
+    Object query() throws SQLException {
+        String sql = createQuery();
+        Connection conn = getDatasource().getConnection();
+        PreparedStatement pstmt = conn.prepareStatement(sql);
+        setValues(pstmt);
+        ResultSet rs = executeQuery(pstmt);
+
+        try (conn; pstmt; rs) {
+            log.debug("query : {}", sql);
+
+            if (rs.next()) {
+                return mapRow(rs);
+            }
+            return null;
         }
     }
